Refactor search.js to be a module
