+++ .gitlab-ci.yml
@@ -38,8 +38,6 @@ build:
       - lizmap/www/assets/js/filter.js.map
       - lizmap/www/assets/js/map.js
       - lizmap/www/assets/js/map.js.map
-      - lizmap/www/assets/js/search.js
-      - lizmap/www/assets/js/search.js.map
       - lizmap/www/assets/js/switcher-layers-actions.js
       - lizmap/www/assets/js/switcher-layers-actions.js.map
       - lizmap/www/assets/js/timemanager.js
